Fixes a bug when CropImageView.onSaveInstanceState is called multiple times
When onSaveInstanceState was called multiple times each Parcelable
result had a reference to the same BitmapUtils.RECT in its
CROP_WINDOW_RECT field. As a result if before serializing that is
changed (by selecting another image for example) it is changed for all
parcelables and it is impossible to restore the previous state.