Scroll the main composer view to the bottom in onMeasure (#119)
When the software keyboard appears in the composer, we want to show the
bottom of the card instead of the bottom of the edit field. To do that
a conditional logic switch is added to the onMeasure method of the main
composer scroll view. When the post text field is focused, the switch is
turned on and calls to onMeasure will trigger scrolling to the bottom.