make stats more usable in grafana
Dimensions work differently than I expected; if you do not specify what
values the dimension can take, it will only show events for which the
dimension was null. Similarly, you cannot specify a filter to include
null values for a dimension. The conclusion is that all dimensions
should always be non-null.