remove Preconditions stack frames from top of crash stack traces
Crashlytics groups by whatever is at the top of the stack. When multiple
functions crash in the same helper (such as Preconditions.checkNotNull)
then the top frame is the same in all cases and so they are all grouped
together. This change removes the top Preconditions frames so that the
different call sites will be grouped correctly.