fix reply container race condition and refactor
Previously for post replies we would end up simultaneously loading both
the post and message. Since for post replies the message is null and the
post needs to be fetched from db, in most cases the message reply loads
first clearing the reply container and then after that the post reply
loads properly. The bug where the post reply sometimes did not appear
occurred when those two events happened in the less likely ordering. Now
all of the reply info is merged into a single ComputableLiveData and
we will only attempt to load one of the two reply types.