fix issue on Android 11 on Google devices with drawing in wrong places
Repro steps: Wait until all syncs have finished (need to look at logs).
Give focus to an edit text and then lock the screen. Wait about 5 seconds
(look for onStop() to be called) and then unlock the device. Observe
that about 50% of the time the top of the window is no longer visible
(sometimes part of the toolbar, sometimes even more) or sometimes the
bottom of the window will be partially obscured as well. I have tried
this on most of my devices, and it only repros on physical Google devices
on Android 11.

I tested throughout the app and was able to repro this only in the
locations changed