improve invite intent handling
The impetus for this was that sharing via FB Messenger was non-functional.
It turned out that the main reason this happened was that we were using
FB's SMS intent but that intent only works when Messenger is the
default SMS app. An additional reason on some newer devices was
that the chooser will only look at the first 2 or 3 elements
of the initial intents extra; if Messenger is installed but WA
and W4B are earlier in the list, then Messenger will not be displayed
in the chooser on API 30 even if the others aren't installed.

For the first problem, we check whether Messenger is the default
SMS app; if it is, then we can use the phone number intent to make
sure it goes to the correct person, and if not, then we use a separate
intent to allow them to choose the recipient. Only one of these two
(regular message and SMS intents) will be displayed. For the second
problem, we prioritize installed and enabled apps in the list.

NOTE: Once we target API 30 we will need to add queries to our
Android Manifest to find out which apps are installed and enabled.