prevent reading internal files
This is a security vulnerability which would have required chaining to
get anything meaningful -- while another app could trick ours into reading
our internal files, we would fail to calculate the dimensions of non-media
and therefore not upload the files. The closest an attacker could get is
being able to leak photos or videos that the user has access to (i.e.
that one of their friends shared), though this requires knowledge of the
file name. We now refuse to read files in internal storage.