Fix group going towards bottom of list when post with 10 items sent
In the getPosts SQL query we go through the post itself and then
the individual media items, up to the limit. When we tried to
fetch the last group post, we provided a limit of 10, so a post
with a few media items would be fine, but if the post had 10
media items the cursor would have 10 + 1 = 11 > 10 rows in it
and so we wouldn't complete a single post, thus returning empty.