Prevent delayed disconnect in connection (#454)
Currently a call to disconnect() shuts down the packet handlers and then
adds disconnectInBackground() to the execution queue. If the queue is
not empty, then it's highly likely that the disconnect call will be
delayed, and potentially additional disconnect calls will be placed in
the queue. This change clears the queue before executing the
background portion.