Refresh and reuse download urls for media that is already uploaded (#456)
* Refresh and reuse download urls for media that is already uploaded

Compute hash for the pre encrypted file and check if we have already
uploaded a file with such a hash. If we did, attempt to reuse the
downloadUrl for that media. If the downloadUrl is valid, do not upload
the media again, but use the url + the old encryption key and hash.

* Use the DEC prefix for hashes of unencrypted media files

* Rename hash parameter

* Fix rebase errors