Migrate speed dial off JCenter
This caused an NPE because the updated speed dial library uses an updated
version of the support library and now HomeFragment's onCreateView is
being called prior to MainActivity's onCreate so that the Fragment is
getting the support action bar before the Activity has set it. We address
this by delaying the addition of the scroll listener.

java.lang.NullPointerException: Attempt to invoke virtual method 'float androidx.appcompat.app.ActionBar.getElevation()' on a null object reference
        at com.halloapp.widget.ActionBarShadowOnScrollListener.updateElevation(ActionBarShadowOnScrollListener.java:62)
        at com.halloapp.widget.ActionBarShadowOnScrollListener.onScrolled(ActionBarShadowOnScrollListener.java:48)
        at com.halloapp.ui.home.HomeFragment$2.onScrolled(HomeFragment.java:255)