Clear reload request on use in PostsDataSource
It is my suspicion that the issue we were seeing with a group post
showing up under the groups tab but not on the home feed was due to
us continuing to use Long.MAX_VALUE as the last loaded key indefinitely
after scrolling to the top. Discussed with Dima and we agreed that
clearing the keyTimestamp field was safe to try as a fix and that
longer-term we should find a more elegant solution, likely removing
the field.