Correctly handle ContactSyncError
Error type would force us into the exception path way. This
is more of a bandaid that redirects the Iq to the result pathway if it's a different type of error
Iq. Im thinking we probably want to refactor so there's exceptions
as well as error IQs. It feels weird forcing all error iqs through as exceptions