Fix ViewHolderWithLifecycle
Setting state to Destroyed when the view is detached (going off screen)
causes the observer to no longer observe state. By setting to created,
we stop getting updates (while off screen) then get updates again
when we are on screen.

We should stop observing when the view gets recycled. I don't think this
is guaranteed to be called once we hit onDestroy in the parent activity.
But they should all get garbage collected