Fixes for call activity recreation. (#937)
* Fixes for call activity recreation.

If you are in a call and rotate the phone and the activity gets
recreated we were trying to start a new call. I added some code to
reject the new call request if we are already in a call. I think a
better fix might be to check if we are already in a call and then don't
do some of the actions.

* Now using the callViewModel State in the CallActivity.onCreate to
decide if any action should be taken.