Fix ConcurrentModificationException on media upload
Introduced in f0f94bdb484eb389bb94d4428b4cb07a26fc14e8

The app would crash the first time trying to upload media because
we would attempt to add an observer to ContentDbObservers when
the UploadMediaTask class was being loaded due to notifyPostAdded
calling MainConnectionObserver's onPostAdded while iterating over
the observer list. The synchronization did not prevent this from
occurring because we used a monitor lock which is reentrant and
we ran all of this on the same thread.