Fix for CallActivity no buttons UI (#1097)
* Fix for CallActivity no buttons UI

If we receive an incoming call and then while the user clicks the
notificaiton to open the call activity we can receive the end of the
call. This makes it possible to have the call activity in the idle state
where there are no buttons. Instead just close the activity.
fixes #1090