Fix post comment indicator not updating
Introduced in 6e42ff41d9f67c8bf49cbba667d00f4ee08ef6aa
We were adjusting numbers down when binding to a post that was further
down in the list than the invite card but not adjusting them back up
from the PostDiffer which only knows about item positions and not the
adjusted positions or the invite card index. This was causing us to
rebind at an index off by one if it was after the invite card, resulting
in indefinitely stale data. It was more likely to show up on older posts
because of the timestamp refresher triggering more often for recent posts.