Fix replies to voice messages showing original not found
Only happened if a message has already been sent in this chat.
We reset the media indices to -1 on message send, making us not notice
that the replied-to message had media, so we did not store text for
the reply preview, and subsequently the null text caused us to conclude
that the replied-to message did not exist.