Fix videos getting reencoded causing hash mismatches for receivers
What was happening was TransferPendingItemsTask was reading
the media transfer state from the db just prior to the finishing
UploadMediaTask writing it out, meaning that when the pending
task started a new upload task the media transfer state was
stale. We now read the transfer state from disk when starting
the upload to ensure it is fresh.