Fix reloading causing crash for post list differ
I made an invalid assumption that the differ will
only have a single change applied ("e.g. 10 elements removed)
but it actually has all of the batch changes all applied. This breaks
reloading as it's two remove operations (one for tail, one for head)
Which causes the index out of bounds issues