Do not allow state changes if teardown key matches
Previously we would receive the first message and teardown the
session, then if we got another message from the previous session
we would process it blindly, skipping inbound keys until we got
the wrongly-generated key with correct index, which would definitely
cause decryption failure, but when we detect the failure we did
not clean up the state because the teardown key matched, meaning
that on a subsequent message our state would be bad. A common way
this manifested was that we would try to use one of the skipped
keys we wrongly generated while processing the second message to
decrypt the resent message, which again guaranteed failure.
With this change we check whether the teardown key matches the
inbound ephemeral key before making any state changes so that when
the rerequested message is received we are in a clean slate.