Fix share button text going onto a second line on small devices
This fixes the issue for all but the smallest devices. We measure
how much space each requires and then use that proportion for the
weights. Other approaches I attempted caused the touch targets to
shrink which looked bad when tapped or made assumptions about the
relative length of the translations of 'comment' and 'share'.