Add lint detector for using % at end of string
This was happening reasonably often on RTL strings where it looked
like the string started with '%s' and then had the RTL translation
but it was actually the RTL translation followed by 's%', which
looks identical. It is unclear why this is not caught by the standard
Android lint rules.