Fix crash where outbound moments missing media if send tapped too quickly
If the user tapped send before the media had been loaded, we would still
save the post to disk without any media attached. This would crash when
we tried to encode the moment to bytes or when we tried to display the
moment. For this reason about 60% of the crashes happened when we tried
to transfer pending items and the other 40% happened when we tried to
display our own moment on the home screen. This is resolved by disabling
the send button until the content is loaded. There is only one user
still hitting this crash and it will resolve when the moment in question
expires within 24 hours.