Remove h265 codec usage
We are not able to play videos on the web if they were encoded using
h265. Originally we planned to reencode videos when performing an
external share to work around this, but now with the web client
happening this model would not make sense since clients would
potentially have to reencode and reupload for inbound as well as
outbound content, resulting in much more duplicated work.
This commit defaults us to h264 now.