Fix deadlock in CallManager (#1363)
We were not acquiring locks in a consistent order, which
resulted in deadlock. In particular, sometimes we acquired
a lock on this and then on timer, and other times we
acquired a lock on timer and then on this. This commit
switches to only locking on this to prevent deadlock.