Fix list of users needing sender state never updating
The first time the group was set up it would function properly, but
every time someone reregistered or was added they would get added to
the list and never removed even if we sent content to them, which was
inefficient. This properly removes them the next time we share the
sender state with them.