Fix EasyPermissions permanently denied usage
This was annoying me, but voice permissions would always take you
to the settings screen initially. This is because the way this works
is:

isPermanentlyDenied == shouldShowRequestPermissionRationale

The way you're supposed to check if its permanently denied is:

on request being denied (permanently denied denies instantly) you
check if shouldShowRequestPermissionRationale is false.