Remove transfer state check for now
If the user was allowed to read the plaintext (at this point only
external users) then we might not find the tombstone so we would
attempt an insert instead of an update which would fail with a
constraint violation, resulting in us not updating the failure reason
and other columns. The only way I know of that we wouldn't find
the tombstone is if the content item had media attached and all
media was downloaded prior to the content being received again,
meaning that the transfer state was updated to TRANSFERRED_YES.
This means that our stats have sometimes been counting success on
a rerequest as a failure due to the column not getting updated.
This removed condition in the where clause will be added back
once both group and home feed crypto are fully rolled out.