Do not send out home feed post on encryption failure
We already do this for all group feed content. We will still send
out home feed comments that fail to encrypt for now because
there are some historical errors that caused the comment key to
be omitted when commenting from iOS and otherwise the only way
the comment key should be missing is if the post failed to decrypt,
in which case we soon will be showing the tombstone instead of
the post which does not have a comment button.