Fix InitialSyncActivity waiting on id of a skipped work item
Introduced in 38d76199105c93cec8ebb4db96c08dd49044787f and cherry-picked
into 1.3.6 in 2f0c5c1ad3578772494331f796d86ccf1989450c

If a partial sync was initiated while a previous sync was still running,
we would update the lastSyncRequestId but the work item would be dropped
since we would use an existing work policy of KEEP, so the activity would
never move on to the feed, waiting on a work item that will never finish.
A similar problem could happen in the other parts of the app where we
do a full contact refresh, but with much smaller impact. We fix this
problem by only keeping track of ids for full syncs since they will always
be run and we only ever care about tracking full syncs.