Respond to rerequests of retracted feed content using msg instead of iq
This should not change any user-visible behavior because it being
retracted in the database means that a retract iq will be sent out at
some point, but this makes it easier to understand what happened
when reading through the logs -- right now the server drops duplicated
retracts, so it looked like the client was acking these rerequests
without handling them. The server will not drop the msg, so the
retract will show up again after the rerequest on the other client.