Implement post reactions
This commit does not include future proof parsing or handling of
rerequests; both will come in future commits. The post_reactions
server prop determines whether the post footer shows the large
reply privately button or a small reply privately button, a react
button, and the list of reactions. The server prop does not hide
reactions from the original poster on the avatars layout or the
seen by screen if they are available.