Fix reactions memory leaks
We were creating content observers in the view holders but when
the view holders were garbage collected ContentDb would still have
a reference to the observer. We also were leaking ReactionLoaders
in the comment view holders. This change makes it less likely that
a reaction will fail to update properly but does not completely
prevent that from happening.