Fix redelivery of content items resulting in duplicated media
We had the incorrect assumption that we would only ever correctly
decrypt a given content item a single time, but in bad network
conditions the server could deliver the same content item multiple
times and the latter deliveries will fail to decrypt, resulting
in a rerequest which causes the sender to reencrypt such that we
will have more than one successful decrypt for the same item.
We address this by checking the transfer state before handling
other items like media.