Set up flavors so we can move to building a new app
Some libraries appear not to support product flavors with different
application IDs. I managed to work around most issues, but I believe
that Huawei's Push Kit integration is broken by this commit, and I
think this is acceptable given the low number of users we saw that
had Huawei push tokens but not Firebase push tokens.