Fix race condition in connection causing packets to be dropped
This was happening particularly often at startup. While it was
possible for this to happen for HalloApp, it was more likely for it
to happen for Katchup because we send more IQs at startup, and it
also was more impactful for Katchup because HalloApp followed the
fully offline model, whereas Katchup needs to fetch fresh data
on startup.

The race condition itself occurred because there was a period of
time between the socket connector reporting that a socket had been
successfully connected and the connection actually making use of
this socket. If another packet was queued up during this time, it
was possible for a new connection attempt to be initiated, and
then when this second attempt succeeded the new socket would be
reported to the connection, causing all of the connection infra
to be restarted. In particular, the init() functions called on
PacketWriter, IqRouter, and MsgRouter would cause packets to be
dropped. Additionally, the already-running PacketReader could try to
read from the new socket before the authentication has been completed
due to the socket being switched out.

This has been fixed by having the connection inform the socket
connector when it is done handling the onSocketConnected callback
and the socket connector refusing to start any more connection
attempts until it has heard back from the connection that the
socket was handled.