Fix registration skipping name setup
In HalloApp we had provided the phone number as the name in
79ec6fc0b9a03054103b9c0e749f829d4ced462c when we were changing
the registration flow to ask for name later rather than earlier.
This was causing the name to already be set locally even though
it had not been set on the server. In this change we no longer
provide the phone number as the name since the server seems
to accept registration without a name now. To prevent the app
always going back to the phone input screen we have to recognize
that a user can be registered without having a name set.