Reduce notification vector precision
We were hitting STRING_TOO_LARGE when trying to post the notification
because we went over the ~32k limit, causing the notification to be
aborted. Since the evenOdd method is used to cut the text out of the
larger circle, the usual strategy of splitting the path into multiple
smaller paths under the limit did not work. Instead we reduce the
precision of the vector by rounding all numbers to at most two decimal
places. Specifically, sed 's/\(\.[0-9][0-9]\)[0-9]*/\1/g' was used
to perform the rounding.