Fix onboarding steps sometimes getting skipped
If the MainActivity had not been destroyed while the onboarding
activities were opened, then we would come back without checking
the registration status again. It took me a long time to figure
this out because most of my devices have "don't keep activities"
enabled, so MainActivity was always getting destroyed right after
we left it, and thus the registration status would be recomputed
when we reopened the activity.