Fix avatars sometimes missing
The original AvatarLoader was built on top of the assumption that you
would only ever show avatars of users in your address book since the
server would not provide the avatar id to random users for privacy
reasons. For some places in the app this was not a problem because
the API returned the avatar id and we were able to directly pass this
into the avatar loader (i.e. following suggestions), but for parts
that did not have easy access to the known avatar id (i.e. the feed)
the avatar would not be loaded if the user was not in your contacts.
Note that such users could appear in either following feed or public
feed since we moved to the following model from the contacts model.
The fix is to collect the avatar id whenever the server provides it
and put this data in the avatars table for usage elsewhere in the app.