Fix memory issues playing selfies in feed
I decided to try limiting buffering based on this post on Stack Overflow:
https://stackoverflow.com/a/64900778/11817085
Experimentally it does indeed appear that ExoPlayer is loading extra
copies of the video into memory when it is shorter than the amount it
is trying to buffer. Specifically, it appears that it wants to load
5000ms into memory and since our selfies are only 1000ms, it ends up
loading FIVE COPIES of the selfie into memory for playback. Adding
these limits resulted in the 80% decrease in memory usage you might
expect based on this understanding.