Update toolbar and header for album prompts
Alisa wanted this behavior to specifically match how Google Maps
handles scrolling where the content fades in on the toolbar when you
scroll past the header and fades back out whenever the header is
visible. I also confirmed that we want to keep the 'it is now or
later' text for all other prompt types.