Update manifest so that we do not intercept privacy policy below API 31
This also affected the Terms of Service, but our app update was
rejected due to the privacy policy not loading. Basically, it turns
out that pathAdvancedPattern as an attribute was only introduced
in API 31, so on API 30 and lower the attribute was ignored, resulting
in all intents related to katchup.com being intercepted, which caused
the privacy policy and terms of service to not load.

The way that our urls work is that user profiles are accessed by going
to katchup.com/<username> and the web directory is under katchup.com/w/,
and since usernames must be at least 3 characters long and are not
allowed to contain '/', these can be distinguished. Unfortunately no
other attribute for intent filters allows enough specificity to
encode this distinction, so we have to manually list out enough
prefixes that are guaranteed not to be /w/. We do this by having
paths that start with a letter other than w and paths that start with
w and then another character which is valid as the 2nd charater in
a username ([a-z_\.0-9] according to the spec)