Fix posts showing in wrong month and empty month in archive
I was wrong in 379ebc090534f9bccfb739a426f71c8ce5c7a3eb -- the empty
error was actually due to the posts not being returned in order of
increasing notification timestamps. This commit sorts the posts so
that they are in the correct order and restores the initial timestamp
calculations so that the posts show up correctly.