Fix onboarding looping issue when no internet
If we cannot reach the server, we were getting stuck in a loop where
registration fails on the permissions onboarding activity, but we
still would try to go to the next step by taking the user to the
MainActivity, but MainActivity would see we haven't registered yet
and so take us back to the permissions activity which would attempt
registration again. Instead we now show an error when this
registration fails.