Fix crash when posting due to missing notification date
Introduced in 8ea9da1206313d0e45998e30978d078a2081da7a

When the user upgrades from a version before we save the moment
notification date to after, posting would be non-functional until they
receive the next moment notification since the date will be null and
protobufs do not allow setting null, causing a crash. Since we try
to resend pending items on each connect, the app would repeatedly
crash and as such be unusable. We work around this by first checking
whether the date is null before setting it on the protobuf.