Fix double opening of activities from MainActivity race condition
It was possible for the registration status result to get sent both
when the view model was instantiated and when the ComputableLiveData
was invalidated in onStart, resulting in the observer being run twice.
I don't think postValue checks whether the new value is equivalent to
the old value, but regardless CheckResult does not override equals
and the new object will be at a different memory address.